+++ R/utils.R
@@ -1,2 +1,2 @@
 # Declarar variables globales para evitar NOTEs en R CMD CHECK
-utils::globalVariables(c("arid_humans", "arid_animals", "arid_plants", ".data"))
+utils::globalVariables(c("arid_humans", "arid_animals", "arid_plants", "arid_c14", ".data"))